Fix issue when cutting strings with a certain number of characters
In Rust, len() returns the number of bytes and not the number of characters, so it is not correct to use this function to trim a string to a certain number of characters.

Instead of doing that, we convert the string slice of the description into a character iterator and with it we create a 250-iteration for loop; Inside the loop we are 'filling' a String (A UTF-8–encoded, growable string) with each of the UTF-8 characters.